Took a double sledgehammer to G2 and it worked

+++ g2core/board/ArduinoDue/hardware.h
@@ -118,7 +118,7 @@ using Motate::OutputPin;
 
 #define FREQUENCY_DDA       100000UL        // Hz step frequency. Interrupts actually fire at 2x (200 KHz)
 #define FREQUENCY_DWELL     1000UL
-#define MIN_SEGMENT_MS ((float)1.0) 
+#define MIN_SEGMENT_MS ((double)1.0) 
 
 #define PLANNER_QUEUE_SIZE (48)
 #define SECONDARY_QUEUE_SIZE (10)
