+++ g2core/device/max31865/max31865.h
@@ -65,13 +65,13 @@ struct MAX31865 final {
 
     // The resulting value is relative to the pullup resistance
     // To return the correct resistance, we need the pullup value
-    float _pullup_resistance;
+    double _pullup_resistance;
 
     // Constructor - this is the only time we directly use the SBIBus
     template <typename SPIBus_t, typename chipSelect_t>
     MAX31865(SPIBus_t &spi_bus,
              const chipSelect_t &_cs,
-             float pullup_resistance = 430,         // 430 is the value used on the Adafruit breakout
+             double pullup_resistance = 430,         // 430 is the value used on the Adafruit breakout
              bool is_three_pin = false,
              bool fifty_hz = true
              )
@@ -93,7 +93,7 @@ struct MAX31865 final {
              std::function<void(bool)> &&_interrupt,
              SPIBus_t &spi_bus,
              const chipSelect_t &_cs,
-             float pullup_resistance = 430,         // 430 is the value used on the Adafruit breakout
+             double pullup_resistance = 430,         // 430 is the value used on the Adafruit breakout
              bool is_three_pin = false,
              bool fifty_hz = true
              )
@@ -401,10 +401,10 @@ struct MAX31865 final {
         return _rtd_value;
     };
 
-    float getPullupResistance() {
+    double getPullupResistance() {
         return _pullup_resistance;
     }
-    void setPullupResistance(const float r) {
+    void setPullupResistance(const double r) {
         _pullup_resistance = r;
     }
 
@@ -416,36 +416,36 @@ struct MAX31865 final {
 //    int32_t getBottom() {
 //        return 0;
 //    };
-//    float getBottomVoltage() {
+//    double getBottomVoltage() {
 //        return 0;
 //    };
 //    int32_t getTop() {
 //        return 32767;
 //    };
-//    float getTopVoltage() {
+//    double getTopVoltage() {
 //        return _vref;
 //    };
 
-    void setVoltageRange(const float vref,
-                         const float min_expected = 0,
-                         const float max_expected = -1,
-                         const float ideal_steps = 1)
+    void setVoltageRange(const double vref,
+                         const double min_expected = 0,
+                         const double max_expected = -1,
+                         const double ideal_steps = 1)
     {
 //        _vref = vref;
 
         // All of the rest are ignored, but here for compatibility of interface
     };
-    float getVoltage() {
-        float r = getRaw();
+    double getVoltage() {
+        double r = getRaw();
         if (r < 0) {
             return r*1000.0;
         }
         return ((r*_pullup_resistance)/32768.0);
     };
-    operator float() { return getVoltage(); };
+    operator double() { return getVoltage(); };
 
-    float getResistance() {
-        float r = getRaw();
+    double getResistance() {
+        double r = getRaw();
         if (r < 0) {
             return r*1000.0;
         }
@@ -512,14 +512,14 @@ public:
         return true;
     };
 
-    float getValue() override
+    double getValue() override
     {
         if (enabled != IO_ENABLED) {
             return 0;
         }
         return pin.getVoltage();
     };
-    float getResistance() override
+    double getResistance() override
     {
         if (enabled != IO_ENABLED) {
             return -1;
@@ -554,14 +554,14 @@ public:
         return false;
     };
 
-    float getParameter(const uint8_t p) override
+    double getParameter(const uint8_t p) override
     {
         if (p == 0) {
             return pin.getPullupResistance();
         }
         return 0;
     };
-    bool setParameter(const uint8_t p, const float v) override
+    bool setParameter(const uint8_t p, const double v) override
     {
         if (p == 0) {
             pin.setPullupResistance(v);
@@ -599,7 +599,7 @@ public:
     // support function for pin value update interrupt handling
 
     void adc_has_new_value(bool err) {
-//        float raw_adc_value = pin.getRaw();
+//        double raw_adc_value = pin.getRaw();
 //        history.add_sample(raw_adc_value);
     };
 };
